UT for Entitlement.WithClock
